Extract mark/common
