+++ test/compile/mark/common.test.ts
@@ -1,0 +1,39 @@
+/* tslint:disable:quotemark */
+
+import {assert} from 'chai';
+import {parseUnitModel} from '../../util';
+import {applyColorAndOpacity} from '../../../src/compile/mark/common';
+
+describe('compile/mark/common', () => {
+  describe('applyColorAndOpacity()', function() {
+    it('opacity should be mapped to a field if specified', function() {
+      const model = parseUnitModel({
+        "mark": "bar",
+        "encoding": {
+          "y": {"type": "quantitative", "field": 'US_Gross', "aggregate": "sum", "axis": true},
+          "opacity": {"field": "US_Gross", "type": "quantitative"}
+        },
+        "data": {"url": "data/movies.json"}
+      });
+
+      let p: any = {};
+      applyColorAndOpacity(p, model);
+      assert.deepEqual(p.opacity.field, 'US_Gross');
+    });
+
+    it('opacity should be mapped to a value if specified', function() {
+      const model = parseUnitModel({
+        "mark": "bar",
+        "encoding": {
+          "y": {"type": "quantitative", "field": 'US_Gross', "aggregate": "sum", "axis": true},
+          "opacity": {"value": 0.5}
+        },
+        "data": {"url": "data/movies.json"}
+      });
+
+      let p: any = {};
+      applyColorAndOpacity(p, model);
+      assert.deepEqual(p.opacity.value, 0.5);
+    });
+  });
+});
